+++ CHANGELOG.md
@@ -7,10 +7,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-### Fixed
+### Changed
+* bumped: metrics version in rockspec
+* bumped: cartridge version in rockspec
+* Role-based model support has been implemented in CRUD (https://jira.vk.team/browse/TNTP-2177)
 
+### Added
+* Validation of `bucket_id`. Invalid values now raise `BucketIDError`
+  before routing. [TNTP-3536](https://jira.vk.team/browse/TNTP-3536)
+
+### Fixed
+* Fixed compatibility with cartridge `2.16.0` ([TNTP-3598](https://jira.vk.team/browse/TNTP-3598))
+* `crud.schema` no longer returns TCF system space `_cdc_state`.
 * `crud.schema` no longer returns system space `_gc_consumers` with Tarantool 3.2+.
+* `crud.schema` no longer returns `tt` system space `_tt_migrations`.
 * Tests of `schema` with Tarantool 3.2+.
+* Fixed bad error handling for method `call.single`
+* Added support for working in 3.1 with data from 2.11, previously there was an error
+due to the inability to find the replicasets by name https://github.com/tarantool/crud-ee/issues/16
 
 ## [1.5.2] - 20-05-24
 
@@ -22,7 +36,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [1.5.1] - 27-04-24
 
-### Added
+### Changed
 * Clarify license in the rockspec (#434).
 
 ## [1.5.0] - 03-04-24